+++ common/src/main/java/com/daqem/jobsplus/client/options/JobsScreenOptions.java
@@ -14,17 +14,17 @@ public class JobsScreenOptions {
     private final List<Job> jobs;
     private final List<Job> preformingJobs;
     private final List<Job> notPreformingJobs;
-    private int coins;
+    private double coins;
     private Job selectedJob;
     private LeftTab selectedLeftTab;
     private RightTab selectedRightTab;
     private SideTab selectedSideTab;
 
-    public JobsScreenOptions(List<Job> jobs, int coins) {
+    public JobsScreenOptions(List<Job> jobs, double coins) {
         this(jobs, coins, jobs.stream().findFirst().orElse(null), LeftTab.ALL, RightTab.INFO, null);
     }
 
-    public JobsScreenOptions(List<Job> jobs, int coins, Job selectedJob, LeftTab selectedLeftTab, RightTab selectedRightTab, @Nullable SideTab selectedSideTab) {
+    public JobsScreenOptions(List<Job> jobs, double coins, Job selectedJob, LeftTab selectedLeftTab, RightTab selectedRightTab, @Nullable SideTab selectedSideTab) {
         this.jobs = jobs.stream()
                 .sorted(Comparator.comparing(Job::getLevel).reversed()
                         .thenComparing(job -> job.getJobInstance().getName().getString()))
@@ -50,7 +50,7 @@ public class JobsScreenOptions {
         return notPreformingJobs;
     }
 
-    public int getCoins() {
+    public double getCoins() {
         return coins;
     }
 
@@ -70,7 +70,7 @@ public class JobsScreenOptions {
         return selectedSideTab;
     }
 
-    public void setCoins(int coins) {
+    public void setCoins(double coins) {
         this.coins = coins;
     }
