+++ common/src/main/java/com/daqem/jobsplus/networking/s2c/ClientboundOpenJobsScreenPacket.java
@@ -21,7 +21,7 @@ import java.util.List;
 public class ClientboundOpenJobsScreenPacket implements CustomPacketPayload {
 
     private final List<Job> jobs;
-    private final int coins;
+    private final double coins;
 
     public static final StreamCodec<RegistryFriendlyByteBuf, ClientboundOpenJobsScreenPacket> STREAM_CODEC = new StreamCodec<>() {
         @Override
@@ -32,18 +32,18 @@ public class ClientboundOpenJobsScreenPacket implements CustomPacketPayload {
         @Override
         public void encode(RegistryFriendlyByteBuf buf, ClientboundOpenJobsScreenPacket packet) {
             buf.writeCollection(packet.jobs, Job.Serializer::toNetwork);
-            buf.writeInt(packet.coins);
+            buf.writeDouble(packet.coins);
         }
     };
 
-    public ClientboundOpenJobsScreenPacket(List<Job> jobs, int coins) {
+    public ClientboundOpenJobsScreenPacket(List<Job> jobs, double coins) {
         this.jobs = jobs;
         this.coins = coins;
     }
 
     public ClientboundOpenJobsScreenPacket(RegistryFriendlyByteBuf friendlyByteBuf) {
         this.jobs = friendlyByteBuf.readList(friendlyByteBuf1 -> Job.Serializer.fromNetwork(friendlyByteBuf1, null));
-        this.coins = friendlyByteBuf.readInt();
+        this.coins = friendlyByteBuf.readDouble();
     }
 
     @Override
